pyproject.toml
[tool.poetry]
name = "trading_bot"
version = "0.1.0"
description = "Trading bot for Deribit"
authors = ["Nikita Maslov nmaslov.test@gmail.com"]
license = "non-commercial"

[tool.poetry.dependencies]
python = "^3.9"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
